+++ core/src/components/accordion/test/standalone/accordion.e2e.ts
@@ -14,6 +14,8 @@ test.describe('accordion: standalone', () => {
   test('should not have visual regressions', async ({ page }) => {
     await page.goto(`/src/components/accordion/test/standalone`);
 
+    await page.waitForChanges();
+
     expect(await page.screenshot({ fullPage: true })).toMatchSnapshot(
       `accordion-standalone-${page.getSnapshotSettings()}.png`
     );