+++ core/src/components/input/test/masking/input.e2e.ts
@@ -1,0 +1,20 @@
+import { expect } from '@playwright/test';
+import { test } from '@utils/test/playwright';
+
+test.describe('input: masking', () => {
+  test.beforeEach(async ({ page }) => {
+    await page.goto('/src/components/input/test/masking');
+  });
+
+  test('should filter out spaces', async ({ page, skip }) => {
+    skip.rtl();
+
+    const input = page.locator('#inputTrimmed');
+
+    await input.click();
+
+    await page.keyboard.type('S p a c e s');
+
+    await expect(input).toHaveJSProperty('value', 'Spaces');
+  });
+});