+++ core/src/components/refresher/refresher.tsx
@@ -10,7 +10,7 @@ import {
   ION_CONTENT_ELEMENT_SELECTOR,
   printIonContentErrorMsg,
 } from '../../utils/content';
-import { clamp, getElementRoot, raf, transitionEndAsync } from '../../utils/helpers';
+import { clamp, componentOnReady, getElementRoot, raf, transitionEndAsync } from '../../utils/helpers';
 import { hapticImpact } from '../../utils/native/haptic';
 
 import {
@@ -443,53 +443,48 @@ export class Refresher implements ComponentInterface {
       console.error('Make sure you use: <ion-refresher slot="fixed">');
       return;
     }
-
     const contentEl = this.el.closest(ION_CONTENT_ELEMENT_SELECTOR);
     if (!contentEl) {
       printIonContentErrorMsg(this.el);
       return;
     }
-
-    const customScrollTarget = contentEl.querySelector(ION_CONTENT_CLASS_SELECTOR);
-
     /**
-     * Query the custom scroll target (if available), first. In refresher implementations,
-     * the ion-refresher element will always be a direct child of ion-content (slot="fixed"). By
-     * querying the custom scroll target first and falling back to the ion-content element,
-     * the correct scroll element will be returned by the implementation.
+     * Waits for the content to be ready before querying the scroll
+     * or the background content element.
      */
-    this.scrollEl = await getScrollElement(customScrollTarget ?? contentEl);
+    componentOnReady(contentEl, async () => {
+      const customScrollTarget = contentEl.querySelector(ION_CONTENT_CLASS_SELECTOR);
+      /**
+       * Query the custom scroll target (if available), first. In refresher implementations,
+       * the ion-refresher element will always be a direct child of ion-content (slot="fixed"). By
+       * querying the custom scroll target first and falling back to the ion-content element,
+       * the correct scroll element will be returned by the implementation.
+       */
+      this.scrollEl = await getScrollElement(customScrollTarget ?? contentEl);
+      /**
+       * Query the background content element from the host ion-content element directly.
+       */
+      this.backgroundContentEl = await contentEl.getBackgroundElement();
 
-    /**
-     * Query the host `ion-content` directly (if it is available), to use its
-     * inner #background-content has the target. Otherwise fallback to the
-     * custom scroll target host.
-     *
-     * This makes it so that implementers do not need to re-create the background content
-     * element and styles.
-     */
-    this.backgroundContentEl = getElementRoot(contentEl ?? customScrollTarget).querySelector(
-      '#background-content'
-    ) as HTMLElement;
+      if (await shouldUseNativeRefresher(this.el, getIonMode(this))) {
+        this.setupNativeRefresher(contentEl);
+      } else {
+        this.gesture = (await import('../../utils/gesture')).createGesture({
+          el: contentEl,
+          gestureName: 'refresher',
+          gesturePriority: 31,
+          direction: 'y',
+          threshold: 20,
+          passive: false,
+          canStart: () => this.canStart(),
+          onStart: () => this.onStart(),
+          onMove: (ev) => this.onMove(ev),
+          onEnd: () => this.onEnd(),
+        });
 
-    if (await shouldUseNativeRefresher(this.el, getIonMode(this))) {
-      this.setupNativeRefresher(contentEl);
-    } else {
-      this.gesture = (await import('../../utils/gesture')).createGesture({
-        el: contentEl,
-        gestureName: 'refresher',
-        gesturePriority: 31,
-        direction: 'y',
-        threshold: 20,
-        passive: false,
-        canStart: () => this.canStart(),
-        onStart: () => this.onStart(),
-        onMove: (ev) => this.onMove(ev),
-        onEnd: () => this.onEnd(),
-      });
-
-      this.disabledChanged();
-    }
+        this.disabledChanged();
+      }
+    });
   }
 
   disconnectedCallback() {
@@ -582,7 +577,7 @@ export class Refresher implements ComponentInterface {
     // best to do any DOM read/writes only when absolutely necessary
     // if multi-touch then get out immediately
     const ev = detail.event as TouchEvent;
-    if (ev.touches && ev.touches.length > 1) {
+    if (ev.touches !== undefined && ev.touches.length > 1) {
       return;
     }